HMW-1004 Cache S3 files on server for better memory performance.
