+++ app/server/app/app.js
@@ -8,6 +8,7 @@ const { checkClientRouteExists } = require('./middleware');
 const { getEnvironment } = require('./server/utilities/environment');
 const logger = require('./server/utilities/logger');
 const { getS3Config } = require('./server/utilities/s3');
+const updateConfigCache = require('./tasks/updateConfigCache');
 const updateGlossary = require('./tasks/updateGlossary');
 const updateUsgsSiteTypes = require('./tasks/updateUsgsSiteTypes');
 const log = logger.logger;
@@ -83,8 +84,7 @@ if (process.env.GLOSSARY_AUTH) {
 if (process.env.USGS_API_KEY) {
   log.info('USGS_API_KEY environmental variable found, continuing.');
 } else {
-  let msg =
-    'USGS_API_KEY environmental variable NOT set, exiting system.';
+  let msg = 'USGS_API_KEY environmental variable NOT set, exiting system.';
   log.error(msg);
   //process.exit();
   throw new Error('Missing Configuration');
@@ -190,6 +190,15 @@ if (isLocal || process.env.CF_INSTANCE_INDEX === '0') {
     { scheduled: true },
   );
 }
+
+updateConfigCache(app.get('s3BucketUrl'));
+cron.schedule(
+  '*/15 * * * *',
+  () => {
+    updateConfigCache(app.get('s3BucketUrl'));
+  },
+  { scheduled: true },
+);
 
 /****************************************************************
  Setup server and routes
